Refactor linux/GPUMeter into platform independent GPUMeter
