+++ darwin/DarwinMachine.c
@@ -125,3 +125,9 @@ bool Machine_isCPUonline(const Machine* host, unsigned int id) {
 
    return true;
 }
+
+double Machine_updateGpuUsage(Machine* super) {
+   /* Not supported yet */
+   (void)super;
+   return -1;
+}
